Address review notes: reuse find_column_name(), merge tests into sync_schema_tests.cpp
- drop the duplicated column-expression resolution from base_table_definition
  and reuse the existing internal::find_column_name() overload set by moving
  is_column_unique() to storage_impl.h as a free function over db_objects
- move the sync_schema matrix tests into the existing tests/sync_schema_tests.cpp
  instead of a separate file
