+++ dev/storage_impl.h
@@ -123,4 +123,33 @@ namespace sqlite_orm::internal {
         return &std::get<ColIdx>(table.elements).name;
     }
 #endif
+
+    /**
+     *  Checks whether the column with the specified name has a column-level `UNIQUE` constraint
+     *  or is contained in a table-level `UNIQUE` constraint of the given table.
+     */
+    template<class Table, class DBOs, satisfies<is_db_objects, DBOs> = true>
+    bool is_column_unique(const DBOs& dbObjects, const Table& table, const std::string& name) {
+        bool result = false;
+        table.for_each_column([&result, &name](auto& column) {
+            if (column.name == name && column.template is_template<unique_t>()) {
+                result = true;
+            }
+        });
+        if (!result) {
+            using elements_type = elements_type_t<Table>;
+            using unique_index_sequence =
+                filter_tuple_sequence_t<elements_type, check_if_is_template<unique_t>::template fn>;
+            iterate_tuple(table.elements, unique_index_sequence{}, [&dbObjects, &result, &name](auto& uniqueConstraint) {
+                iterate_tuple(uniqueConstraint.columns, [&dbObjects, &result, &name](auto& columnExpression) {
+                    if (const std::string* columnName = find_column_name(dbObjects, columnExpression)) {
+                        if (*columnName == name) {
+                            result = true;
+                        }
+                    }
+                });
+            });
+        }
+        return result;
+    }
 }
